Move blob uri parser function to Parser
