+++ neuro-sdk/src/neuro_sdk/buckets.py
@@ -40,16 +40,13 @@ from neuro_sdk.file_filter import (
     translate,
 )
 from neuro_sdk.file_utils import FileSystem, FileTransferer, LocalFS
-from neuro_sdk.url_utils import (
-    _extract_path,
-    normalize_blob_path_uri,
-    normalize_local_path_uri,
-)
+from neuro_sdk.url_utils import _extract_path, normalize_local_path_uri
 from neuro_sdk.utils import AsyncContextManager
 
 from .config import Config
 from .core import _Core
 from .errors import NDJSONError, ResourceNotFound
+from .parser import Parser
 from .utils import NoPublicConstructor, asyncgeneratorcontextmanager
 
 if sys.version_info >= (3, 7):
@@ -505,9 +502,10 @@ class PersistentBucketCredentials:
 
 
 class Buckets(metaclass=NoPublicConstructor):
-    def __init__(self, core: _Core, config: Config) -> None:
+    def __init__(self, core: _Core, config: Config, parser: Parser) -> None:
         self._core = core
         self._config = config
+        self._parser = parser
 
     def _parse_bucket_payload(self, payload: Mapping[str, Any]) -> Bucket:
         return Bucket(
@@ -618,22 +616,6 @@ class Buckets(metaclass=NoPublicConstructor):
         async with self._get_provider(bucket_name, cluster_name) as provider:
             yield BucketFS(provider)
 
-    def split_blob_uri(self, uri: URL) -> Tuple[str, str, str]:
-        uri = normalize_blob_path_uri(
-            uri, self._config.username, self._config.cluster_name
-        )
-        cluster_name = uri.host
-        assert cluster_name
-        parts = uri.path.lstrip("/").split("/", 2)
-        if len(parts) == 1:
-            raise ValueError(f"Blob uri doesn't contain bucket name: {uri}")
-        if len(parts) == 3:
-            _, bucket_id, key = parts
-        else:
-            _, bucket_id = parts
-            key = ""
-        return cluster_name, bucket_id, key
-
     # Low level operations
 
     async def head_blob(
@@ -680,7 +662,7 @@ class Buckets(metaclass=NoPublicConstructor):
         recursive: bool = False,
         limit: Optional[int] = None,
     ) -> AsyncIterator[BucketEntry]:
-        cluster_name, bucket_name, key = self.split_blob_uri(uri)
+        cluster_name, bucket_name, key = self._parser.split_blob_uri(uri)
         async with self._get_provider(bucket_name, cluster_name) as provider:
             async with provider.list_blobs(key, recursive=recursive, limit=limit) as it:
                 async for entry in it:
@@ -688,7 +670,7 @@ class Buckets(metaclass=NoPublicConstructor):
 
     @asyncgeneratorcontextmanager
     async def glob_blobs(self, uri: URL) -> AsyncIterator[BucketEntry]:
-        cluster_name, bucket_name, key = self.split_blob_uri(uri)
+        cluster_name, bucket_name, key = self._parser.split_blob_uri(uri)
         if _has_magic(bucket_name):
             raise ValueError(
                 "You can not glob on bucket names. Please provide name explicitly."
@@ -764,7 +746,7 @@ class Buckets(metaclass=NoPublicConstructor):
         progress: Optional[AbstractFileProgress] = None,
     ) -> None:
         src = normalize_local_path_uri(src)
-        cluster_name, bucket_name, key = self.split_blob_uri(dst)
+        cluster_name, bucket_name, key = self._parser.split_blob_uri(dst)
         async with self._get_bucket_fs(bucket_name, cluster_name) as bucket_fs:
             transferer = FileTransferer(LocalFS(), bucket_fs)
             await transferer.transfer_file(
@@ -783,7 +765,7 @@ class Buckets(metaclass=NoPublicConstructor):
         continue_: bool = False,
         progress: Optional[AbstractFileProgress] = None,
     ) -> None:
-        cluster_name, bucket_name, key = self.split_blob_uri(src)
+        cluster_name, bucket_name, key = self._parser.split_blob_uri(src)
         dst = normalize_local_path_uri(dst)
         async with self._get_bucket_fs(bucket_name, cluster_name) as bucket_fs:
             transferer = FileTransferer(bucket_fs, LocalFS())
@@ -806,7 +788,7 @@ class Buckets(metaclass=NoPublicConstructor):
         progress: Optional[AbstractRecursiveFileProgress] = None,
     ) -> None:
         src = normalize_local_path_uri(src)
-        cluster_name, bucket_name, key = self.split_blob_uri(dst)
+        cluster_name, bucket_name, key = self._parser.split_blob_uri(dst)
         async with self._get_bucket_fs(bucket_name, cluster_name) as bucket_fs:
             transferer = FileTransferer(LocalFS(), bucket_fs)
             await transferer.transfer_dir(
@@ -828,7 +810,7 @@ class Buckets(metaclass=NoPublicConstructor):
         filter: Optional[AsyncFilterFunc] = None,
         progress: Optional[AbstractRecursiveFileProgress] = None,
     ) -> None:
-        cluster_name, bucket_name, key = self.split_blob_uri(src)
+        cluster_name, bucket_name, key = self._parser.split_blob_uri(src)
         dst = normalize_local_path_uri(dst)
         async with self._get_bucket_fs(bucket_name, cluster_name) as bucket_fs:
             transferer = FileTransferer(bucket_fs, LocalFS())
@@ -842,7 +824,7 @@ class Buckets(metaclass=NoPublicConstructor):
             )
 
     async def blob_is_dir(self, uri: URL) -> bool:
-        cluster_name, bucket_name, key = self.split_blob_uri(uri)
+        cluster_name, bucket_name, key = self._parser.split_blob_uri(uri)
         if key.endswith("/"):
             return True
         async with self._get_bucket_fs(bucket_name, cluster_name) as bucket_fs:
@@ -855,7 +837,7 @@ class Buckets(metaclass=NoPublicConstructor):
         recursive: bool = False,
         progress: Optional[AbstractDeleteProgress] = None,
     ) -> None:
-        cluster_name, bucket_name, key = self.split_blob_uri(uri)
+        cluster_name, bucket_name, key = self._parser.split_blob_uri(uri)
         async with self._get_bucket_fs(bucket_name, cluster_name) as bucket_fs:
             await file_utils.rm(bucket_fs, PurePosixPath(key), recursive, progress)
